server: tests: do not cleanup in CI environment

+++ server/.jest/auth.setup.js
@@ -10,7 +10,9 @@ process.argv.push('--rundir', temporaryRuntimeDirectory);
 process.argv.push('--noauth', 'false');
 
 afterAll(() => {
-  // TODO: This leads to test flakiness caused by ENOENT error
-  // NeDB provides no method to close database connection
-  fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+  if (process.env.CI !== 'true') {
+    // TODO: This leads to test flakiness caused by ENOENT error
+    // NeDB provides no method to close database connection
+    fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+  }
 });
