+++ server/.jest/test.setup.js
@@ -38,9 +38,11 @@ process.argv.push('--allowedpath', temporaryRuntimeDirectory);
 
 afterAll((done) => {
   rTorrentProcess.on('close', () => {
-    // TODO: This leads to test flakiness caused by ENOENT error
-    // NeDB provides no method to close database connection
-    fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+    if (process.env.CI !== 'true') {
+      // TODO: This leads to test flakiness caused by ENOENT error
+      // NeDB provides no method to close database connection
+      fs.rmdirSync(temporaryRuntimeDirectory, {recursive: true});
+    }
     done();
   });
